refactor: [165] Extracted utils into a file, removed tabular
