+++ fenn/__init__.py
@@ -1,3 +1,19 @@
+from fenn.agents import Flow, Node, RAGNode
+from fenn.agents.llm import LLMClient
 from fenn.app import Fenn
+from fenn.nn import Checkpoint, ClassificationTrainer, RegressionTrainer
+from fenn.nn.trainers import LoRATrainer
+from fenn.reproducibility import set_seed
 
-__all__ = ["Fenn"]
+__all__ = [
+    "Fenn",
+    "Flow",
+    "Node",
+    "RAGNode",
+    "LLMClient",
+    "set_seed",
+    "Checkpoint",
+    "ClassificationTrainer",
+    "LoRATrainer",
+    "RegressionTrainer",
+]
